+++ app/Http/Controllers/InsumoController.php
@@ -6,19 +6,13 @@ use PDF;
 use App\Insumo;
 use Illuminate\Http\Request;
 
-/**
- * Class InsumoController
- * @package App\Http\Controllers
- */
 class InsumoController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
+        if(!Auth::check()){
+            return redirect('/login');
+        }
         $insumos = Insumo::paginate();
 
         return view('insumo.index', compact('insumos'))
@@ -35,23 +29,16 @@ class InsumoController extends Controller
         return $pdf->stream();
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
+        if(!Auth::check()){
+            return redirect('/login');
+        }
+
         $insumo = new Insumo();
         return view('insumo.create', compact('insumo'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         request()->validate(Insumo::$rules);
@@ -62,39 +49,28 @@ class InsumoController extends Controller
             ->with('success', 'Insumo creado correctamente.');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
+        if(!Auth::check()){
+            return redirect('/login');
+        }
+
         $insumo = Insumo::find($id);
 
         return view('insumo.show', compact('insumo'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
+        if(!Auth::check()){
+            return redirect('/login');
+        }
+
         $insumo = Insumo::find($id);
 
         return view('insumo.edit', compact('insumo'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request $request
-     * @param  Insumo $insumo
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, Insumo $insumo)
     {
         request()->validate(Insumo::$rules);
@@ -105,13 +81,11 @@ class InsumoController extends Controller
             ->with('success', 'Insumo updated successfully');
     }
 
-    /**
-     * @param int $id
-     * @return \Illuminate\Http\RedirectResponse
-     * @throws \Exception
-     */
     public function destroy($id)
     {
+        if(!Auth::check() || Auth::user()->cargo != 'ADMIN'){
+            return redirect('/login');
+        }
         $insumo = Insumo::find($id)->delete();
 
         return redirect()->route('insumos.index')
