+++ app/Http/Controllers/InsumosProductoController.php
@@ -2,24 +2,19 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
 use App\InsumosProducto;
 use App\Producto;
 use App\Insumo;
 use Illuminate\Http\Request;
 
-/**
- * Class InsumosProductoController
- * @package App\Http\Controllers
- */
 class InsumosProductoController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create(Request $req)
     {
+        if(!Auth::check() || (Auth::user()->cargo != "ADMIN" && Auth::user()->cargo != "ALMACENISTA")){
+            return redirect('/login');
+        }
         $insumosProducto = new InsumosProducto();
         $producto = Producto::find($req->query('id'));
         $insumos = Insumo::whereNotIn('id', $producto->insumos->pluck('id'))->get();
@@ -27,12 +22,6 @@ class InsumosProductoController extends Controller
         return view('insumos_producto.create', compact('insumosProducto', 'producto', 'insumos'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         request()->validate(InsumosProducto::$rules);
@@ -43,12 +32,6 @@ class InsumosProductoController extends Controller
             ->with('success', 'Añadido insumo necesario.');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         $insumosProducto = InsumosProducto::find($id);
@@ -56,14 +39,11 @@ class InsumosProductoController extends Controller
         return view('insumos-producto.show', compact('insumosProducto'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
+        if(!Auth::check() || (Auth::user()->cargo != "ADMIN" && Auth::user()->cargo != "ALMACENISTA")){
+            return redirect('/login');
+        }
         $insumosProducto = InsumosProducto::find($id);
 
         $producto = Producto::find($insumosProducto->producto_id);
@@ -72,13 +52,6 @@ class InsumosProductoController extends Controller
         return view('insumos_producto.edit', compact('insumosProducto', 'producto', 'insumos'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request $request
-     * @param  InsumosProducto $insumosProducto
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, InsumosProducto $insumosProducto)
     {
         request()->validate(InsumosProducto::$rules);
@@ -89,13 +62,11 @@ class InsumosProductoController extends Controller
             ->with('success', 'Insumo necesario actualizado correctamente.');
     }
 
-    /**
-     * @param int $id
-     * @return \Illuminate\Http\RedirectResponse
-     * @throws \Exception
-     */
     public function destroy($id)
     {
+        if(!Auth::check() || (Auth::user()->cargo != "ADMIN" )){
+            return redirect('/login');
+        }
         $insumosProducto = InsumosProducto::find($id);
         $idProducto = $insumosProducto->producto_id;
         $insumosProducto->delete();
