+++ app/Http/Controllers/ProveedoraController.php
@@ -2,45 +2,32 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
 use App\Proveedora;
 use Illuminate\Http\Request;
 
-/**
- * Class ProveedoraController
- * @package App\Http\Controllers
- */
 class ProveedoraController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
+        if(!Auth::check() || (Auth::user()->cargo != "ADMIN" && Auth::user()->cargo != "GESTOR DE COMPRAS")){
+            return redirect('/login');
+        }
         $proveedoras = Proveedora::paginate();
 
         return view('proveedora.index', compact('proveedoras'))
             ->with('i', (request()->input('page', 1) - 1) * $proveedoras->perPage());
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
+        if(!Auth::check() || (Auth::user()->cargo != "ADMIN" && Auth::user()->cargo != "GESTOR DE COMPRAS")){
+            return redirect('/login');
+        }
         $proveedora = new Proveedora();
         return view('proveedora.create', compact('proveedora'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         request()->validate(Proveedora::$rules);
@@ -48,42 +35,29 @@ class ProveedoraController extends Controller
         $proveedora = Proveedora::create($request->all());
 
         return redirect()->route('proveedoras.index')
-            ->with('success', 'Proveedora created successfully.');
+            ->with('success', 'Proveedora creada correctamente.');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
+        if(!Auth::check() || (Auth::user()->cargo != "ADMIN" && Auth::user()->cargo != "GESTOR DE COMPRAS")){
+            return redirect('/login');
+        }
         $proveedora = Proveedora::find($id);
 
         return view('proveedora.show', compact('proveedora'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
+        if(!Auth::check() || (Auth::user()->cargo != "ADMIN" && Auth::user()->cargo != "GESTOR DE COMPRAS")){
+            return redirect('/login');
+        }
         $proveedora = Proveedora::find($id);
 
         return view('proveedora.edit', compact('proveedora'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request $request
-     * @param  Proveedora $proveedora
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, Proveedora $proveedora)
     {
         request()->validate(Proveedora::$rules);
@@ -91,19 +65,17 @@ class ProveedoraController extends Controller
         $proveedora->update($request->all());
 
         return redirect()->route('proveedoras.index')
-            ->with('success', 'Proveedora updated successfully');
+            ->with('success', 'Proveedora actualizada correctamente.');
     }
 
-    /**
-     * @param int $id
-     * @return \Illuminate\Http\RedirectResponse
-     * @throws \Exception
-     */
     public function destroy($id)
     {
+        if(!Auth::check() || (Auth::user()->cargo != "ADMIN")){
+            return redirect('/login');
+        }
         $proveedora = Proveedora::find($id)->delete();
 
         return redirect()->route('proveedoras.index')
-            ->with('success', 'Proveedora deleted successfully');
+            ->with('success', 'Proveedora eliminada correctamente.');
     }
 }
